Deadband-type EMA Noise Filter for Throttle Input
Implemented EMA Filter with variable Alpha based on a defined Noise Threshold. As the rate of change approaches the threshold, we have more confidence in intentional movement and use a larger alpha, up to an instant response.

This most notably helps to eliminate noise and "nervousness" when sustaining a constant throttle position.