Wifi State Management for Battery Savings
Wifi is now managed such that it is turned on when sending data, then we turn it off once we get data back from the receiver, and it remains off until the next data_send. This should result in SIGNIFICANTLY better battery life, especially in Battery Saver mode where the data_send rate is much more limited when unnecessary.